refactor: 요구사항 반영 리팩토링(템플릿)
